1. codelounge-admin-server/requirements.txt
flask 
pyrebase4
python-dotenv
# pkg_resources


# pip install setuptools pyrebase4 flask